requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
alembic==1.11.1
anyio==3.7.0
apache-airflow==2.5.1
apache-airflow-providers-common-sql==1.5.1
apache-airflow-providers-ftp==3.4.1
apache-airflow-providers-http==4.4.1
apache-airflow-providers-imap==3.2.1
apache-airflow-providers-sqlite==3.4.1
apispec==3.3.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
boto3==1.26.154
botocore==1.29.154
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
cloudpickle==2.2.1
cmake==3.26.3
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.3.15
cryptography==41.0.1
dask==2023.5.0
Deprecated==1.2.14
dill==0.3.6
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.2.5
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
flask-session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
fsspec==2023.5.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
lazy-loader==0.2
lazy-object-proxy==1.9.0
linkify-it-py==2.0.2
lit==16.0.5.post0
locket==1.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.2
partd==1.4.0
pathspec==0.9.0
pendulum==2.1.2
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
prison==0.2.1
psutil==5.9.5
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
s3transfer==0.6.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typing-extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
